DeploymentService interfaces refactors making the class hierarchy flatten and making responsibility of interfaces more narrow - DeploymentService interface was merged with DeploymentServiceImpl class - ProcessStateService which was used by ProcessService and ProcessesResources was extracted into the separate file and renamed into ProcessStateProvider - ProcessingTypeDeploymentService was renamed into ProcessingTypeActionService, getDeployedScenarios method was removed from the interface - it was already handled by AllDeployedScenarioService - Part of the DeploymentService interface used by ProcessingTypeActionService (former ProcessingTypeDeploymentService) was extracted into the new ActionService interface - AllDeployedScenarioService was renamed into ProcessingTypeDeployedScenariosProvider
