+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/DefaultProcessingTypeDeployedScenariosProvider.scala
@@ -18,14 +18,14 @@ import scala.util.{Failure, Success}
 
 // This class is extracted from DeploymentService to avoid a cyclic dependency:
 // DeploymentService -> EmbeddedDeploymentManager -> ... -> DeploymentService.getDeployedScenarios
-class AllDeployedScenarioService(
+class DefaultProcessingTypeDeployedScenariosProvider(
     processRepository: FetchingProcessRepository[DB],
     dbioRunner: DBIOActionRunner,
     scenarioResolver: ScenarioResolver,
     processingType: ProcessingType
-) extends LazyLogging {
+) extends ProcessingTypeDeployedScenariosProvider with LazyLogging {
 
-  def getDeployedScenarios(implicit ec: ExecutionContext): Future[List[DeployedScenarioData]] = {
+  override def getDeployedScenarios(implicit ec: ExecutionContext): Future[List[DeployedScenarioData]] = {
     implicit val userFetchingDataFromRepository: LoggedUser = NussknackerInternalUser.instance
     for {
       deployedProcesses <- {
@@ -65,12 +65,12 @@ class AllDeployedScenarioService(
 
 }
 
-object AllDeployedScenarioService {
+object DefaultProcessingTypeDeployedScenariosProvider {
 
   // This factory method prepare objects that are also prepared by AkkaHttpBasedRouteProvider
   // but without dependency to ModelData - it necessary to avoid a cyclic dependency:
   // DeploymentService -> EmbeddedDeploymentManager -> ... -> DeploymentService.getDeployedScenarios
-  def apply(dbRef: DbRef, processingType: ProcessingType)(implicit ec: ExecutionContext): AllDeployedScenarioService = {
+  def apply(dbRef: DbRef, processingType: ProcessingType)(implicit ec: ExecutionContext): ProcessingTypeDeployedScenariosProvider = {
     val dbioRunner = DBIOActionRunner(dbRef)
     val dumbModelInfoProvier = ProcessingTypeDataProvider.withEmptyCombinedData(
       Map(processingType -> ValueWithRestriction.anyUser(Map.empty[String, String]))
@@ -78,7 +78,7 @@ object AllDeployedScenarioService {
     val actionRepository        = new DbProcessActionRepository(dbRef, dumbModelInfoProvier)
     val processRepository       = DBFetchingProcessRepository.create(dbRef, actionRepository)
     val futureProcessRepository = DBFetchingProcessRepository.createFutureRepository(dbRef, actionRepository)
-    new AllDeployedScenarioService(
+    new DefaultProcessingTypeDeployedScenariosProvider(
       processRepository,
       dbioRunner,
       new ScenarioResolver(
