--- designer/server/src/main/scala/pl/touk/nussknacker/ui/process/deployment/DefaultProcessingTypeDeploymentService.scala
@@ -1,27 +1,0 @@
-package pl.touk.nussknacker.ui.process.deployment
-
-import pl.touk.nussknacker.engine.api.deployment.{
-  DeployedScenarioData,
-  ProcessAction,
-  ProcessActionId,
-  ProcessingTypeDeploymentService
-}
-import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessingType}
-
-import scala.concurrent.{ExecutionContext, Future}
-
-class DefaultProcessingTypeDeploymentService(
-    processingType: ProcessingType,
-    deploymentService: DeploymentService,
-    allDeploymentsService: AllDeployedScenarioService
-) extends ProcessingTypeDeploymentService {
-  override def getDeployedScenarios(implicit ec: ExecutionContext): Future[List[DeployedScenarioData]] =
-    allDeploymentsService.getDeployedScenarios
-
-  override def markActionExecutionFinished(actionId: ProcessActionId)(implicit ec: ExecutionContext): Future[Boolean] =
-    deploymentService.markActionExecutionFinished(processingType, actionId)
-
-  override def getLastStateAction(processId: ProcessId)(implicit ec: ExecutionContext): Future[Option[ProcessAction]] =
-    deploymentService.getLastStateAction(processingType, processId)
-
-}
