+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/AkkaHttpBasedRouteProvider.scala
@@ -15,6 +15,7 @@ import pl.touk.nussknacker.engine.api.component.{
   DesignerWideComponentId,
   EmptyAdditionalUIConfigProviderFactory
 }
+import pl.touk.nussknacker.engine.api.deployment.ProcessingTypeDeployedScenariosProvider
 import pl.touk.nussknacker.engine.api.process.ProcessingType
 import pl.touk.nussknacker.engine.compile.ProcessValidator
 import pl.touk.nussknacker.engine.definition.test.ModelDataTestInfoProvider
@@ -89,14 +90,14 @@ class AkkaHttpBasedRouteProvider(
       featureTogglesConfig = FeatureTogglesConfig.create(resolvedConfig)
       _                    = logger.info(s"Designer config loaded: \nfeatureTogglesConfig: $featureTogglesConfig")
       countsReporter <- createCountsReporter(featureTogglesConfig, environment, sttpBackend)
-      deploymentServiceSupplier  = new DelayedInitDeploymentServiceSupplier
+      actionServiceSupplier      = new DelayedInitActionServiceSupplier
       additionalUIConfigProvider = createAdditionalUIConfigProvider(resolvedConfig, sttpBackend)
       typeToConfig <- prepareProcessingTypeData(
         config,
         processingTypeDataStateFactory,
         additionalUIConfigProvider,
-        deploymentServiceSupplier,
-        AllDeployedScenarioService(dbRef, _),
+        actionServiceSupplier,
+        DefaultProcessingTypeDeployedScenariosProvider(dbRef, _),
         sttpBackend,
       )
     } yield {
@@ -167,7 +168,7 @@ class AkkaHttpBasedRouteProvider(
           futureProcessRepository
         )
 
-      val deploymentService = new DeploymentServiceImpl(
+      val deploymentService = new DeploymentService(
         dmDispatcher,
         processRepository,
         actionRepository,
@@ -180,7 +181,7 @@ class AkkaHttpBasedRouteProvider(
       )
       deploymentService.invalidateInProgressActions()
 
-      deploymentServiceSupplier.set(deploymentService)
+      actionServiceSupplier.set(deploymentService)
 
       // we need to reload processing type data after deployment service creation to make sure that it will be done using
       // correct classloader and that won't cause further delays during handling requests
@@ -322,7 +323,7 @@ class AkkaHttpBasedRouteProvider(
         val routes = List(
           new ProcessesResources(
             processService = processService,
-            deploymentService = deploymentService,
+            processStateService = deploymentService,
             processToolbarService = configProcessToolbarService,
             processAuthorizer = processAuthorizer,
             processChangeListener = processChangeListener
@@ -529,8 +530,8 @@ class AkkaHttpBasedRouteProvider(
       designerConfig: ConfigWithUnresolvedVersion,
       processingTypeDataStateFactory: ProcessingTypeDataStateFactory,
       additionalUIConfigProvider: AdditionalUIConfigProvider,
-      deploymentServiceSupplier: Supplier[DeploymentService],
-      createAllDeployedScenarioService: ProcessingType => AllDeployedScenarioService,
+      actionServiceProvider: Supplier[ActionService],
+      createDeployedScenariosProvider: ProcessingType => ProcessingTypeDeployedScenariosProvider,
       sttpBackend: SttpBackend[Future, Any],
   ): Resource[IO, ProcessingTypeDataReload] = {
     Resource
@@ -539,10 +540,10 @@ class AkkaHttpBasedRouteProvider(
           new ProcessingTypeDataReload({ () =>
             def getDeploymentManagerDependencies(processingType: ProcessingType) = {
               DeploymentManagerDependencies(
-                new DefaultProcessingTypeDeploymentService(
+                createDeployedScenariosProvider(processingType),
+                new DefaultProcessingTypeActionService(
                   processingType,
-                  deploymentServiceSupplier.get(),
-                  createAllDeployedScenarioService(processingType)
+                  actionServiceProvider.get(),
                 ),
                 system.dispatcher,
                 system,
@@ -595,19 +596,19 @@ class AkkaHttpBasedRouteProvider(
     additionalUIConfigProviderFactory.create(config, sttpBackend)
   }
 
-  private class DelayedInitDeploymentServiceSupplier extends Supplier[DeploymentService] {
-    private val deploymentServiceRef = new AtomicReference[Option[DeploymentService]](None)
+  private class DelayedInitActionServiceSupplier extends Supplier[ActionService] {
+    private val actionServiceRef = new AtomicReference[Option[ActionService]](None)
 
-    override def get(): DeploymentService = {
-      val deploymentService = deploymentServiceRef.get()
-      deploymentService.getOrElse(
+    override def get(): ActionService = {
+      val actionService = actionServiceRef.get()
+      actionService.getOrElse(
         throw new IllegalStateException(
-          "Illegal initialization: DeploymentService should be initialized before ProcessingTypeData"
+          "Illegal initialization: ActionService should be initialized before ProcessingTypeData"
         )
       )
     }
 
-    def set(deploymentService: DeploymentService): Unit = deploymentServiceRef.set(Some(deploymentService))
+    def set(actionService: ActionService): Unit = actionServiceRef.set(Some(actionService))
   }
 
 }
