+++ engine/flink/management/src/it/scala/pl/touk/nussknacker/engine/management/streaming/FlinkStreamingDeploymentManagerProviderHelper.scala
@@ -3,7 +3,11 @@ package pl.touk.nussknacker.engine.management.streaming
 import akka.actor.ActorSystem
 import org.asynchttpclient.DefaultAsyncHttpClientConfig
 import pl.touk.nussknacker.engine.api.component.DesignerWideComponentId
-import pl.touk.nussknacker.engine.api.deployment.{DeploymentManager, ProcessingTypeDeploymentServiceStub}
+import pl.touk.nussknacker.engine.api.deployment.{
+  DeploymentManager,
+  ProcessingTypeActionServiceStub,
+  ProcessingTypeDeployedScenariosProviderStub
+}
 import pl.touk.nussknacker.engine.management.FlinkStreamingDeploymentManagerProvider
 import pl.touk.nussknacker.engine.{
   ConfigWithUnresolvedVersion,
@@ -29,11 +33,11 @@ object FlinkStreamingDeploymentManagerProviderHelper {
         _ => true
       )
     )
-    val actorSystem       = ActorSystem("FlinkStreamingDeploymentManagerProviderHelper")
-    val backend           = AsyncHttpClientFutureBackend.usingConfig(new DefaultAsyncHttpClientConfig.Builder().build())
-    val deploymentService = new ProcessingTypeDeploymentServiceStub(List.empty)
+    val actorSystem = ActorSystem("FlinkStreamingDeploymentManagerProviderHelper")
+    val backend     = AsyncHttpClientFutureBackend.usingConfig(new DefaultAsyncHttpClientConfig.Builder().build())
     val deploymentManagerDependencies = DeploymentManagerDependencies(
-      deploymentService,
+      new ProcessingTypeDeployedScenariosProviderStub(List.empty),
+      new ProcessingTypeActionServiceStub,
       actorSystem.dispatcher,
       actorSystem,
       backend
